+++ src/Authority/Port.php
@@ -3,8 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Authority;
 
-use Innmind\Url\Exception\DomainException;
-
 /**
  * @psalm-immutable
  */
@@ -15,7 +13,7 @@ final class Port
     private function __construct(?int $value)
     {
         if (\is_int($value) && $value < 0) {
-            throw new DomainException((string) $value);
+            throw new \DomainException((string) $value);
         }
 
         $this->value = $value;