+++ src/Authority/UserInformation/User.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Authority\UserInformation;
 
-use Innmind\Url\{
-    Authority\Host,
-    Exception\DomainException,
-};
+use Innmind\Url\Authority\Host;
 use Innmind\Immutable\Str;
 
 /**
@@ -28,7 +25,7 @@ final class User
     public static function of(string $value): self
     {
         if (!Str::of($value)->matches(self::PATTERN)) {
-            throw new DomainException($value);
+            throw new \DomainException($value);
         }
 
         return new self($value);