+++ tests/Authority/PortTest.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Tests\Authority;
 
-use Innmind\Url\{
-    Authority\Port,
-    Exception\DomainException,
-};
+use Innmind\Url\Authority\Port;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class PortTest extends TestCase
@@ -22,7 +19,7 @@ class PortTest extends TestCase
 
     public function testThrowWhenNegativePort()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(\DomainException::class);
         $this->expectExceptionMessage('-1');
 
         Port::of(-1);