+++ tests/Authority/UserInformation/PasswordTest.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Tests\Authority\UserInformation;
 
-use Innmind\Url\{
-    Authority\UserInformation\Password,
-    Exception\DomainException,
-};
+use Innmind\Url\Authority\UserInformation\Password;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class PasswordTest extends TestCase
@@ -21,7 +18,7 @@ class PasswordTest extends TestCase
 
     public function testThrowWhenInvalidPassword()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(\DomainException::class);
 
         Password::of('foo@bar');
     }