+++ tests/Authority/UserInformation/UserTest.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Url\Tests\Authority\UserInformation;
 
-use Innmind\Url\{
-    Authority\UserInformation\User,
-    Exception\DomainException,
-};
+use Innmind\Url\Authority\UserInformation\User;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class UserTest extends TestCase
@@ -21,7 +18,7 @@ class UserTest extends TestCase
 
     public function testThrowWhenInvalidUser()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(\DomainException::class);
         $this->expectExceptionMessage('user@me');
 
         User::of('user@me');